fixed #111: Possibility for ConcurrentModificationExceptions with Forks
Only send variable updates to clients if we are not in a dry-run
situation. If we are, we don't have anything to send anyway, and if the
dry-run is performed in a fork the useless variable update might cause a
corruption of the variable state just in that moment in which the state
is being sent to the fork (which needs that state right after its
dry-run has finished to sync with the master before actually performing
its run).