attempting to fix issue #122
This adds some more synchronization between forks and masters, in an attempt to fix this strange issue. However, additionally, I have implemented a bunch of checks now to catch illegal or unsafe situations in the termination sequence between forks and masters as well that should facilitate the elimination of externally caused race conditions between processes.