fixed big architectural flaw in timesync feature
The master must calculate the internal parameters required for
"timekeepers" running in the forks to calculate the correct test time
instead of just forwarding the request to set a certain time and speed.
If the latter was done, runtime differences between forks will lead to
differences in the time calculated, which is not desired.