fixes issue #151: Fixed non-working result documentations for suites.
Also changed the logic to detect default results vs. named results to
better match the way JavaDoc handles results (which doesn't know named
results, so we had to add such a thing, but it required a mandatory :,
that was not required for params, so that could create quite some
confusion).