Auto-generate fork binding ports in the forked process by default
This fixes issues #177 and #176. It turns around the fork process port
auto-determination: instead of trying out ports on the master process,
then communicating a port to a fork and hope that it is still free when
the fork has finished starting up, the forked process now finds a free
port to bind to and communicates this to the master via a special log
message (STDOUT) that is understood and parsed by the master process.
Implementors of different forking schemes (like distributed ones with
forks on different machines) might have to make minor adjustments,
especially if they chose to not forward stdout streams to the master (in
that case they will now have to take care of communicating the forks'
port back to the master).