Fixed issue #196 thoroughly by adding a separate fork summary result msg
Fork result summaries (= the aggregate numbers for an execution block on
a fork) used to not be transmitted explicitly, but were scraped out of
the normal set list result transmissions by the master. However, this
creates some problems with detecting abort execution exceptions, which
are rooted in the fact that the master cannot easily discern between
sub-suite results returned by the fork and the "top-level" suite result,
because those look identical to the master. Therefore it would be
cleaner to explicitly return suite result numbers via a separate
dedicated remoting message, thus in turn solving all these problems
caused by the old approach.
This however requires a change to the remoting protocol, which is also
used between test runners and the Eclipse plugin. Such changes are not
desired for patch versions, but only for major and minor versions, which
is why issue #196 is now fixed "twice": once on 0.18.0 in the "clean"
way by adding this new message, and once on 0.17.14 by just removing a
condition around a block on the master that creates a "fake" result
summary for a fork in case of an abortion happening on the fork. The
latter also fixes the issue #196 by guaranteeing the report of at least
one exception if an abort execution exception is encountered, but it
results in possibly wrong result numbers being displayed on the console
on the master.