fixed issue #200: unexpected variable assignment
This actually undos a change that was done in a fix for issue #192,
which was later found to be sub-par and was fixed in a different way in
issue #194. However, the change originally done for issue #192 turned
out to be causing a bug, issue #200. This now is the fix for that bug.