fixes #256: Show multiple whitespaces in HTML result
This was not so easy after all. In order to not have the CSS
"white-space" directive destroy the layout with unintended whitespace
found in longer result values that originate from forced line breaks
injected by Saxon due to it obeying a maximum line length of 120
characters, I had to apply another patch to the Saxon transformer. The
line break length is unfortunately hard-coded, thus the only way to get
rid of this "feature" entirely I had to patch it out of the source code.
The diff accomplishing this change is checked in for future reference.